+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- tls 1.3 is used as the default minimum version
+- only strong cipher suites are allowed for tls version less than 1.3
 
 ## [v0.3.0]
 - upgraded to httpaux v0.2.1